Agregando message404

+++ src/main/java/com/alura/foro/controller/CursoController.java
@@ -7,12 +7,14 @@ import com.alura.foro.record.curso.DatosListadoCurso;
 import com.alura.foro.repository.CursoRepository;
 import com.alura.foro.record.curso.DatosRegistroCurso;
 import com.alura.foro.record.curso.DatosRespuestaCurso;
+import com.alura.foro.util.Util;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import java.net.URI;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -53,34 +55,43 @@ public class CursoController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<DatosRespuestaCurso> RetornaDatosCurso(@PathVariable Long id) {
-        Curso curso = cursoRepository.getReferenceById(id);
-        var datosCurso = new DatosRespuestaCurso(
-                curso.getCursoId(),
-                curso.getNombre(),
-                curso.getCategoria()
-        );
-        return ResponseEntity.ok(datosCurso);
+    public ResponseEntity<?> RetornaDatosCurso(@PathVariable Long id) {
+        if (cursoRepository.existsById(id)) {
+            Curso curso = cursoRepository.getReferenceById(id);
+            var datosCurso = new DatosRespuestaCurso(
+                    curso.getCursoId(),
+                    curso.getNombre(),
+                    curso.getCategoria()
+            );
+            return ResponseEntity.ok(datosCurso);
+        }
+        return new ResponseEntity(new Util().message404(), HttpStatus.NOT_FOUND);
     }
 
     @PutMapping
     @Transactional
     public ResponseEntity actualizarCurso(@RequestBody @Valid DatosActualizarCurso datosActualizarCurso) {
-        Curso curso = cursoRepository.getReferenceById(datosActualizarCurso.id());
-        curso.actualizarDatos(datosActualizarCurso);
-        return ResponseEntity.ok(new DatosRespuestaCurso(
-                curso.getCursoId(),
-                curso.getNombre(),
-                curso.getCategoria()
-        ));
+        if (cursoRepository.existsById(datosActualizarCurso.id())) {
+            Curso curso = cursoRepository.getReferenceById(datosActualizarCurso.id());
+            curso.actualizarDatos(datosActualizarCurso);
+            return ResponseEntity.ok(new DatosRespuestaCurso(
+                    curso.getCursoId(),
+                    curso.getNombre(),
+                    curso.getCategoria()
+            ));
+        }
+        return new ResponseEntity(new Util().message404(), HttpStatus.NOT_FOUND);
     }
-    
+
     @DeleteMapping("/{id}")
     @Transactional
-    public ResponseEntity eliminarCurso(@PathVariable  Long id){
-        Curso curso = cursoRepository.getReferenceById(id);
-        cursoRepository.delete(curso);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity eliminarCurso(@PathVariable Long id) {
+        if (cursoRepository.existsById(id)) {
+            Curso curso = cursoRepository.getReferenceById(id);
+            cursoRepository.delete(curso);
+            return ResponseEntity.noContent().build();
+        }
+        return new ResponseEntity(new Util().message404(), HttpStatus.NOT_FOUND);
     }
 
 }
